day 19/requirements.txt
fastapi
uvicorn
python-dotenv
requests
assemblyai
Jinja2
python-multipart
google-generativeai